More portable code

+++ libnspire/src/callback.rs
@@ -1,11 +1,11 @@
-use std::os::raw::{c_ulong, c_void};
+use std::os::raw::{c_void};
 
 pub struct CallbackData<'a>(pub &'a mut dyn FnMut(usize));
 
 impl CallbackData<'_> {
-    pub unsafe extern "C" fn callback(size: c_ulong, data: *mut c_void) {
+    pub unsafe extern "C" fn callback(size: usize, data: *mut c_void) {
         let data = &mut *(data as *mut CallbackData);
-        data.0(size as usize);
+        data.0(size);
     }
     pub fn as_mut_void(&mut self) -> *mut c_void {
         self as *mut CallbackData as *mut c_void
